admin_panel_service/movies_admin/requirements.txt
django==4.2
flake8==6.1
python-dotenv==1.0
django-split-settings==1.2
psycopg2-binary==2.9.7
django-debug-toolbar==3.5
